requirements.txt
colorama
python-slugify
flickrapi
six
Unidecode
urllib3
